flask_api/requirements.txt
flask
langchain
langchain-groq
pydantic
python-dotenv
asyncio
langchain-community